+++ src/core/computedvalue.ts
@@ -90,13 +90,14 @@ export class ComputedValue<T> implements IObservable, IComputedValue<T>, IDeriva
 	 */
 	public get(): T {
 		invariant(!this.isComputing, `Cycle detected in computation ${this.name}`, this.derivation);
-		startBatch();
-		if (globalState.inBatch === 1) {
+		if (globalState.inBatch === 0) {
 			// just for small optimization, can be droped for simplicity
 			// computed called outside of any mobx stuff. batch observing shuold be enough, don't need tracking
 			// because it will never be called again inside this batch
+			startBatch();
 			if (shouldCompute(this))
 				this.value = this.peek();
+			endBatch();
 		} else {
 			reportObserved(this);
 			if (shouldCompute(this))
@@ -104,7 +105,6 @@ export class ComputedValue<T> implements IObservable, IComputedValue<T>, IDeriva
 					propagateChangeConfirmed(this);
 		}
 		const result = this.value!;
-		endBatch();
 
 		if (result instanceof CaughtException)
 			throw result.cause;
@@ -128,10 +128,9 @@ export class ComputedValue<T> implements IObservable, IComputedValue<T>, IDeriva
 	private trackAndCompute(): boolean {
 		if (isSpyEnabled()) {
 			spyReport({
-				object: this,
+				object: this.scope,
 				type: "compute",
-				fn: this.derivation,
-				target: this.scope
+				fn: this.derivation
 			});
 		}
 		const oldValue = this.value;