+++ src/core/transaction.ts
@@ -1,40 +1,30 @@
-import {globalState} from "./globalstate";
-import {runReactions} from "./reaction";
 import {startBatch, endBatch} from "./observable";
-import {isSpyEnabled, spyReportStart, spyReportEnd} from "./spy";
+import {deprecated} from "../utils/utils";
+import {untrackedStart, untrackedEnd} from "./derivation";
 
 /**
+ * @deprecated
  * During a transaction no views are updated until the end of the transaction.
  * The transaction will be run synchronously nonetheless.
+ *
+ * Deprecated to simplify api; transactions offer no real benefit above actions.
+ *
  * @param action a function that updates some reactive state
  * @returns any value that was returned by the 'action' parameter.
  */
-export function transaction<T>(action: () => T, thisArg = undefined, report = true): T {
-	transactionStart(((action as any).name) || "anonymous transaction", thisArg, report);
+export function transaction<T>(action: () => T, thisArg = undefined): T {
+	deprecated("Using `transaction` is deprecated, use `runInAction` or `(@)action` instead.");
+	return runInTransaction.apply(undefined, arguments);
+}
+
+export function runInTransaction<T>(action: () => T, thisArg = undefined): T {
+	// TODO: use execute action!
+	startBatch();
+	const prev = untrackedStart();
 	try {
 		return action.call(thisArg);
 	} finally {
-		transactionEnd(report);
+		untrackedEnd(prev);
+		endBatch();
 	}
-}
-
-export function transactionStart<T>(name: string, thisArg, report: boolean) {
-	startBatch();
-	globalState.inTransaction += 1;
-	if (report && isSpyEnabled()) {
-		spyReportStart({
-			type: "transaction",
-			target: thisArg,
-			name: name
-		});
-	}
-}
-
-export function transactionEnd<T>(report: boolean) {
-	if (--globalState.inTransaction === 0) {
-		runReactions();
-	}
-	if (report && isSpyEnabled())
-		spyReportEnd();
-	endBatch();
 }